renamed forecast & extracted weather code
renamed forecast to WeatherForecast.

extracted most of code in Weather to a separate component called WeatherReport to separate concerns
