+++ src/components/index.ts
@@ -1,9 +1,19 @@
 import Button from './Button';
-import Forecast from './Forecast';
 import Header from './Header';
 import Search from './Search';
 import SearchForm from './SearchForm';
 import SearchHistory from './SearchHistory';
 import Weather from './Weather';
+import WeatherForecast from './WeatherForecast';
+import WeatherReport from './WeatherReport';
 
-export { Button, Forecast, Header, Search, SearchForm, SearchHistory, Weather };
+export {
+  Button,
+  Header,
+  Search,
+  SearchForm,
+  SearchHistory,
+  Weather,
+  WeatherForecast,
+  WeatherReport,
+};
